refactor : StationSteps - 요청과 검증 Steps 로 분리
