+++ src/test/java/nextstep/subway/line/LineAcceptanceTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static nextstep.subway.line.LineRequestSteps.*;
-import static nextstep.subway.line.LineVerificationStep.*;
+import static nextstep.subway.line.LineVerificationSteps.*;
 
 @DisplayName("지하철 노선 관련 기능")
 public class LineAcceptanceTest extends AcceptanceTest {
