+++ src/test/java/nextstep/subway/line/LineVerificationSteps.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class LineVerificationStep {
+public class LineVerificationSteps {
 
     public static void 지하철_노선_생성_됨(ExtractableResponse<Response> response) {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
