+++ src/test/java/nextstep/subway/station/StationAcceptanceTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static nextstep.subway.station.StationSteps.*;
+import static nextstep.subway.station.StationVerificationSteps.*;
 
 @DisplayName("지하철역 관련 기능")
 public class StationAcceptanceTest extends AcceptanceTest {
